MNT/ENH: Add a raw CCSDS packet generator
This refactors out the reading of a byte-stream into a separate
generator function that can yield raw CCSDSPackets with only
a header and the raw bytes. The definitions and parsing logic can
use this generator to yield packets and then parse whatever they
need to after the fact. This enables someone to use this generator
without a packet definition and to investigate what is in the packet
stream without needing to know more detailed information about
how individual packets are laid out internally.
